src/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.2
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
idna==2.5
kombu==4.0.2
olefile==0.44
Pillow==4.1.1
python-resize-image==1.1.11
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
vine==1.1.3
